+++ packages/app-builder-lib/src/options/SnapOptions.ts
@@ -11,8 +11,7 @@ export interface SnapOptions extends CommonLinuxOptions, TargetSpecificOptions {
   /**
    * The custom environment. Defaults to `{"TMPDIR: "$XDG_RUNTIME_DIR"}`. If you set custom, it will be merged with default.
    */
-  readonly environment?: object | null
-  // { [key: string]: string } Our scheme generator cannot validate this type
+  readonly environment?: { [key: string]: string } | null
 
   /**
    * The 78 character long summary. Defaults to [productName](/configuration/configuration#Configuration-productName).
@@ -82,6 +81,12 @@ export interface SnapOptions extends CommonLinuxOptions, TargetSpecificOptions {
    * Whether to use template snap. Defaults to `true` if `stagePackages` not specified.
    */
   readonly useTemplateApp?: boolean
+
+  /**
+   * Whether or not the snap should automatically start on login.
+   * @default false
+   */
+  readonly autoStart?: boolean
 }
 
 export interface PlugDescriptor {